requirements.txt
streamlit==1.37.0
openai==1.37.1
python-dotenv==1.0.1
chromadb==0.5.5
protobuf==3.20.1
pysqlite3-binary==0.5.3
llama-index==0.10.58